+++ assets/javascripts/kitten/config/configured-radium.js
@@ -1,18 +1,6 @@
+// DEPRECATED: This configuration is the Radium default configuration.
+
 import Radium from 'radium'
-import { resolveArrayValue } from 'css-in-js-utils'
-
-// Custom plugin to fix array values.
-const radiumResolveArrayValue = (values) => {
-  Object.keys(values.style).map(
-    (key) => {
-      const stylesByKey = values.style[key]
-
-      if (Array.isArray(stylesByKey)) {
-        values.style[key] = resolveArrayValue(key, stylesByKey)
-      }
-    }
-  )
-}
 
 const config = {
   userAgent: 'all',
@@ -25,7 +13,6 @@ const config = {
     Radium.Plugins.visited,
     Radium.Plugins.removeNestedStyles,
     Radium.Plugins.prefix,
-    radiumResolveArrayValue,
     Radium.Plugins.checkProps,
   ],
 }