+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix : Fix margin on HorizontalNav component.
+
 ## [19.2.0] - 2018-01-22
 
 Feature: