+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Twitter`, `Facebook` and `Linkedin` social button icon words.
+
 ## [19.2.0] - 2018-01-22
 
 Feature: