+++ CHANGELOG.md
@@ -4,10 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `focus()` to `TextInput`.
+
 ## [19.2.0] - 2018-01-22
 
 Feature:
-- Add `focus()` to `TextInput`.
 - Update styles `disabled` prop on `Button` and `ButtonIcon`.
 
 Fixes: